Define ruby dependency for bundler
Previous to this commit, there was no explicit relationship between
Class['bundler'] and Class['robby::ruby'].  This caused bundler to be
installed prior to ruby 1.9 being set up.  That in turn caused all the
gems for the Robby application to be installed in ruby 1.8 path and
therefor unavailable to the application.

This commit creates an explicit dependency between the classes which
resolves the above issues.