Use packages to deploy application
This commit removes vcsrepo as the prefered method to deploy the
Robby application.  Instead, it uses a package to get the files on disk.
Since git is no longer necessary, the deploy key has been removed and
the ssh key for the robby user is no longer being managed.

Further, this commit removes the unicorn resource definition out of the
robby class so it can be better placed in a profile.